+++ pallets/runtime/tests/src/relayer_test.rs
@@ -1,3 +1,4 @@
+use codec::Encode;
 use frame_support::dispatch::{DispatchInfo, Pays, PostDispatchInfo};
 use frame_support::weights::Weight;
 use frame_support::{assert_noop, assert_ok};
@@ -7,16 +8,19 @@ use sp_runtime::traits::{Dispatchable, TransactionExtension, TxBaseImplication};
 use sp_runtime::transaction_validity::TransactionSource;
 use sp_runtime::transaction_validity::{InvalidTransaction, TransactionValidityError};
 use sp_runtime::MultiAddress;
+use sp_runtime::MultiSignature;
 
-use pallet_relayer::Subsidy;
+use pallet_balances::Call as BalancesCall;
+use pallet_relayer::{RelayTxNonces, Subsidy, UniqueCall};
 use polymesh_primitives::constants::currency::POLY;
 use polymesh_primitives::protocol_fee::ProtocolOp;
 use polymesh_primitives::traits::CurrentFeePayer;
 use polymesh_primitives::{AccountId, Balance, Ticker, TransactionError};
-use polymesh_runtime_develop::runtime::{TxFeeHandler, RuntimeCall as DevRuntimeCall};
+use polymesh_runtime_develop::runtime::{RuntimeCall as DevRuntimeCall, TxFeeHandler};
 use polymesh_transaction_payment::Val;
 
-use super::storage::{RuntimeCall, TestStorage, User};
+use super::pips_test::assert_balance;
+use super::storage::{register_keyring_account_with_balance, RuntimeCall, TestStorage, User};
 use super::ExtBuilder;
 
 type Relayer = pallet_relayer::Pallet<TestStorage>;
@@ -674,5 +678,97 @@ fn do_relayer_accept_cdd_and_fees_test() {
             bob.acc()
         ),
         Ok(Some(alice.acc()))
+    );
+}
+
+#[test]
+fn relay_happy_case() {
+    ExtBuilder::default()
+        .build()
+        .execute_with(_relay_happy_case);
+}
+
+fn _relay_happy_case() {
+    let alice = Sr25519Keyring::Alice.to_account_id();
+    let _ = register_keyring_account_with_balance(Sr25519Keyring::Alice, 1_000).unwrap();
+
+    let bob = Sr25519Keyring::Bob.to_account_id();
+    let _ = register_keyring_account_with_balance(Sr25519Keyring::Bob, 1_000).unwrap();
+
+    let charlie = Sr25519Keyring::Charlie.to_account_id();
+    let _ = register_keyring_account_with_balance(Sr25519Keyring::Charlie, 1_000).unwrap();
+
+    // 41 Extra for registering a DID
+    assert_balance(bob.clone(), 1041, 0);
+    assert_balance(charlie.clone(), 1041, 0);
+
+    let origin = RuntimeOrigin::signed(alice);
+    let transaction = UniqueCall::new(
+        RelayTxNonces::<TestStorage>::get(bob.clone()),
+        RuntimeCall::Balances(BalancesCall::transfer_allow_death {
+            dest: charlie.clone().into(),
+            value: 50,
+        }),
+    );
+
+    assert_ok!(Relayer::relay_tx(
+        origin,
+        bob.clone(),
+        Sr25519Keyring::Bob.sign(&transaction.encode()).into(),
+        transaction
+    ));
+
+    assert_balance(bob, 991, 0);
+    assert_balance(charlie, 1_091, 0);
+}
+
+#[test]
+fn relay_unhappy_cases() {
+    ExtBuilder::default()
+        .build()
+        .execute_with(_relay_unhappy_cases);
+}
+
+fn _relay_unhappy_cases() {
+    let alice = Sr25519Keyring::Alice.to_account_id();
+    let _ = register_keyring_account_with_balance(Sr25519Keyring::Alice, 1_000).unwrap();
+
+    let bob = Sr25519Keyring::Bob.to_account_id();
+
+    let charlie = Sr25519Keyring::Charlie.to_account_id();
+
+    let origin = RuntimeOrigin::signed(alice);
+    let transaction = UniqueCall::new(
+        RelayTxNonces::<TestStorage>::get(bob.clone()),
+        RuntimeCall::Balances(BalancesCall::transfer_allow_death {
+            dest: charlie.clone().into(),
+            value: 59,
+        }),
+    );
+
+    let signature = MultiSignature::Sr25519(Default::default());
+    assert_noop!(
+        Relayer::relay_tx(
+            origin.clone(),
+            bob.clone(),
+            signature.clone(),
+            transaction.clone()
+        ),
+        Error::InvalidSignature
+    );
+
+    let _ = register_keyring_account_with_balance(Sr25519Keyring::Bob, 1_000).unwrap();
+
+    let transaction = UniqueCall::new(
+        RelayTxNonces::<TestStorage>::get(bob.clone()) + 1,
+        RuntimeCall::Balances(BalancesCall::transfer_allow_death {
+            dest: charlie.into(),
+            value: 59,
+        }),
+    );
+
+    assert_noop!(
+        Relayer::relay_tx(origin.clone(), bob, signature, transaction),
+        Error::InvalidNonce
     );
 }
