+++ pallets/utility/Cargo.toml
@@ -6,9 +6,7 @@ edition = "2021"
 
 [dependencies]
 # Out Pallets
-pallet-balances = { workspace = true, default-features = false }
 pallet-permissions = { workspace = true, default-features = false }
-pallet-identity = { workspace = true, default-features = false }
 polymesh-transaction-payment = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 
@@ -39,15 +37,12 @@ std = [
 	"sp-std/std",
 	"polymesh-transaction-payment/std",
 	"polymesh-primitives/std",
-	"pallet-balances/std",
-	"pallet-identity/std",
 ]
 
 runtime-benchmarks = [
 	"frame-benchmarking/runtime-benchmarks",
 	"frame-support/runtime-benchmarks",
 	"frame-system/runtime-benchmarks",
-	"pallet-identity/runtime-benchmarks",
 	"polymesh-transaction-payment/runtime-benchmarks",
 ]
 try-runtime = ["frame-support/try-runtime"]
