+++ pallets/weights/src/pallet_relayer.rs
@@ -101,4 +101,24 @@ impl pallet_relayer::WeightInfo for SubstrateWeight {
             .saturating_add(DbWeight::get().reads(1))
             .saturating_add(DbWeight::get().writes(1))
     }
+    // Storage: Identity KeyRecords (r:2 w:0)
+    // Proof Skipped: Identity KeyRecords (max_values: None, max_size: None, mode: Measured)
+    // Storage: Utility Nonces (r:1 w:1)
+    // Proof: Utility Nonces (max_values: None, max_size: Some(48), added: 2523, mode: MaxEncodedLen)
+    // Storage: Timestamp Now (r:1 w:0)
+    // Proof: Timestamp Now (max_values: Some(1), max_size: Some(8), added: 503, mode: MaxEncodedLen)
+    // Storage: Instance2Group ActiveMembers (r:1 w:0)
+    // Proof Skipped: Instance2Group ActiveMembers (max_values: Some(1), max_size: None, mode: Measured)
+    // Storage: Identity Claims (r:2 w:0)
+    // Proof Skipped: Identity Claims (max_values: None, max_size: None, mode: Measured)
+    // Storage: Permissions CurrentPalletName (r:1 w:1)
+    // Proof Skipped: Permissions CurrentPalletName (max_values: Some(1), max_size: None, mode: Measured)
+    // Storage: Permissions CurrentDispatchableName (r:1 w:1)
+    // Proof Skipped: Permissions CurrentDispatchableName (max_values: Some(1), max_size: None, mode: Measured)
+    fn relay_tx() -> Weight {
+        // Minimum execution time: 128_001 nanoseconds.
+        Weight::from_parts(138_117_000, 0)
+            .saturating_add(DbWeight::get().reads(9))
+            .saturating_add(DbWeight::get().writes(3))
+    }
 }
